Temporary fix for using Switchery as a RequireJS module
Switchery uses ComponentJS as its dependency and build mechanism.
There's a compatability issue when a module built with ComponentJS is loaded with dependencies by RequireJS.
The fix changes the standalone dist/switchery.js file directly, which is a bad idea.

We need to find a better solution for this issue.

See https://github.com/componentjs/component/issues/627